+++ packages/web-components/src/components/leadspace/__stories__/leadspace.stories.react.tsx
@@ -92,9 +92,8 @@ const iconOptions = {
   PDF: 'Pdf20',
 };
 
-const Default = ({ parameters }) => {
-  const { alt, buttons, copy, defaultSrc, hasImage, hasVideo, navElements, size, title, type } =
-    parameters?.props?.Leadspace ?? {};
+const Default = args => {
+  const { alt, buttons, copy, defaultSrc, hasImage, hasVideo, navElements, size, title, type } = args?.Leadspace ?? {};
   return (
     <DDSLeadspace size={size} type={type} {...(hasImage ? { alt } : {})}>
       {navElements === navigationOptions[0] ? navigationWithTagGroup : ``}
@@ -298,8 +297,8 @@ MediumWithVideo.story = {
   },
 };
 
-export const Short = ({ parameters }) => {
-  const { alt, defaultSrc, hasImage, hasVideo, navElements, size, title, type } = parameters?.props?.Leadspace ?? {};
+export const Short = args => {
+  const { alt, defaultSrc, hasImage, hasVideo, navElements, size, title, type } = args?.Leadspace ?? {};
   return (
     <DDSLeadspace size={size} type={type} {...(hasImage ? { alt } : {})}>
       {navElements === navigationOptions[0] ? navigationWithTagGroup : ``}