+++ packages/services-store/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.38.0](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/compare/@carbon/ibmdotcom-services-store@1.38.0-rc.0...@carbon/ibmdotcom-services-store@1.38.0) (2022-09-13)
+
+**Note:** Version bump only for package @carbon/ibmdotcom-services-store
+
+
+
+
+
+# [1.38.0-rc.0](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/compare/@carbon/ibmdotcom-services-store@1.37.0...@carbon/ibmdotcom-services-store@1.38.0-rc.0) (2022-08-29)
+
+**Note:** Version bump only for package @carbon/ibmdotcom-services-store
+
+
+
+
+
 # [1.37.0](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/compare/@carbon/ibmdotcom-services-store@1.37.0-rc.1...@carbon/ibmdotcom-services-store@1.37.0) (2022-08-17)
 
 **Note:** Version bump only for package @carbon/ibmdotcom-services-store